+++ lib/chat_web/live/main_live/page/lobby.ex
@@ -36,6 +36,12 @@ defmodule ChatWeb.MainLive.Page.Lobby do
     |> process(&join_approved_requests/1)
   end
 
+  def refresh_rooms_and_users(socket) do
+    socket
+    |> assign_room_list()
+    |> assign_user_list()
+  end
+
   def new_room(%{assigns: %{me: me, monotonic_offset: time_offset}} = socket, name, type)
       when type in [:public, :request, :private, :cargo] do
     {cargo?, type} =