+++ CHANGELOG.md
@@ -18,6 +18,7 @@ straightforward as possible.
 ### Fixed
 - Too long buttons on mobile
 - scroll uploads to top on uploader mount
+- Update users/rooms cache on USB plugging/unplugging
 
 ## 2023-05-12_7e5fcce___2023-05-12_b47aebb